% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{blake3_hash_raw}
\alias{blake3_hash_raw}
\title{Hash a raw vector using BLAKE3}
\usage{
blake3_hash_raw(x, key = NULL)
}
\arguments{
\item{x}{a raw vector}

\item{key}{either NULL or a 32 byte long raw vector}
}
\description{
Hash a raw vector using BLAKE3
}
